+++ src/components/WinTable/index.module.sass
