+++ src/pages/Home/Body/index.js
@@ -1,13 +1,13 @@
 import "./index.css";
-import Profile from "./Profile";
-import WinTable from "./WinTable";
-import LeagueCards from "./LeagueCards";
-import MatchOverview from "./MatchOverview";
+import Profile from "../../../components/Profile";
+import WinTable from "../../../components/WinTable";
+import LeagueCards from "../../../components/LeagueCards";
+import MatchOverview from "../../../components/MatchOverview";
 import {
   useGetPlayerResult,
   useGetSummonerWinRate,
   useGetMatches,
-} from "../../../../query";
+} from "../../../query";
 import { useSelector } from "react-redux";
 
 const Body = () => {
