+++ src/pages/Home/View.js
@@ -1,5 +1,5 @@
-import Header from "./components/Header";
-import Body from "./components/Body";
+import Header from "./Header";
+import Body from "./Body";
 
 const View = () => {
   return (
